Add capability check in the lookup path
In capability mode, only relative lookups are allowed. For *at
syscalls, only  strict relatives lookups from the fd are allowed.

For now, symlinks which go outside the sandbox are not allowed
either, according to freeBSD semantic.

A race can occur with rename, hence we cannot rely on the ncp to
check we don't escape the root, since an attack can use rename out
of the sandbox to attack the sandbox itself.

After a loockup, we check that the ncp is under the sandbox root
by walking upwards from the result ncp.

BUG : in the current implementation, we can leak information from
the hierarchy outside the sandbox : indeed, we only check if we
escape the sandbox at the end of the lookup, hence the lookup can
return a different error like ENOENT before that